Language.Haskell.Tools.Refactor

Description

Defines the API for refactorings

Documentation

module Language.Haskell.Tools.AST.SemaInfoClasses

module Language.Haskell.Tools.Rewrite

module Language.Haskell.Tools.AST.References

module Language.Haskell.Tools.AST.Helpers

module Language.Haskell.Tools.Refactor.Utils.Monadic

module Language.Haskell.Tools.Refactor.Utils.Helpers

module Language.Haskell.Tools.Rewrite.ElementTypes

data UnsupportedExtension

Constructors

UnsupportedExtension String

type TypedModule = Ann UModule IdDom SrcTemplateStage

The final version of our AST, with type infromation added

tryRefactor :: (RealSrcSpan -> Refactoring) -> String -> ModuleName -> IO ()

A quick function to try the refactorings

correctRefactorSpan :: UnnamedModule -> RealSrcSpan -> RealSrcSpan

Adjust the source range to be applied to the refactored module

useFlags :: [String] -> Ghc ([String], DynFlags -> DynFlags)

Set the given flags for the GHC session. Also gives back a change function that you can use to apply the settings to any flags. Prints out errors and warnings

pprWarning :: Warn -> SDoc

pprErr :: Err -> SDoc

reloadPkgDb :: Ghc ()

Reloads the package database based on the session flags

initGhcFlags :: Ghc ()

Initialize GHC flags to default values that support refactoring

initGhcFlagsForTest :: Ghc ()

initGhcFlags' :: Bool -> Bool -> Ghc ()

Sets up basic flags and settings for GHC

useDirs :: [FilePath] -> Ghc ()

Use the given source directories when searching for imported modules

deregisterDirs :: [FilePath] -> Ghc ()

Don't use the given source directories when searching for imported modules

toFileName :: FilePath -> ModuleName -> FilePath

Translates module name and working directory into the name of the file where the given module should be defined

toBootFileName :: FilePath -> ModuleName -> FilePath

Translates module name and working directory into the name of the file where the boot module should be defined

getSourceDir :: ModSummary -> IO FilePath

Get the source directory where the module is located.

getModSumOrig :: ModSummary -> FilePath

Gets the path to the source file of the module.

keyFromMS :: ModSummary -> SourceFileKey

getModSumName :: ModSummary -> String

Gets the module name

loadModuleAST :: FilePath -> ModuleName -> Ghc TypedModule

Load the AST of a module given by the working directory and module name.

loadModule :: FilePath -> ModuleName -> Ghc ModSummary

Load the summary of a module given by the working directory and module name.

parseTyped :: ModSummary -> Ghc TypedModule

Get the typed representation of a Haskell module.

trfProblem :: String -> a

withAlteredDynFlags :: GhcMonad m => (DynFlags -> m DynFlags) -> m a -> m a

Modifies the dynamic flags for performing a ghc task

forceCodeGen :: ModSummary -> ModSummary

Forces the code generation for a given module

codeGenDfs :: DynFlags -> DynFlags

forceAsmGen :: ModSummary -> ModSummary

Forces ASM code generation for a given module

modSumNormalizeFlags :: ModSummary -> ModSummary

Normalizes the flags for a module summary

normalizeFlags :: DynFlags -> DynFlags

Removes all flags that are unintelligable for refactoring

readSrcSpan :: String -> RealSrcSpan

Read a source range from our textual format: line:col-line:col or line:col

readSrcLoc :: String -> RealSrcLoc

Read a source location from our format: line:col

module Language.Haskell.Tools.Refactor.Utils.Lists

module Language.Haskell.Tools.Refactor.Utils.BindingElem

module Language.Haskell.Tools.Refactor.Utils.Indentation

module Language.Haskell.Tools.Refactor.Querying

module Language.Haskell.Tools.Refactor.Refactoring

module Language.Haskell.Tools.Refactor.Utils.Name

module Language.Haskell.Tools.Refactor.Representation

module Language.Haskell.Tools.Refactor.Monad

module Language.Haskell.Tools.Refactor.Utils.Type

module Language.Haskell.Tools.Refactor.Utils.TypeLookup

module Language.Haskell.Tools.Refactor.Utils.NameLookup

data Ann (elem :: * -> * -> *) dom stage

An element of the AST keeping extra information.

class HasSourceInfo e where

Minimal complete definition

srcInfo

Associated Types

type SourceInfoType e :: *

Methods

srcInfo :: Simple Lens e (SourceInfoType e)

class HasRange a where

Extracts or modifies the concrete range corresponding to a given source info. In case of lists and optional elements, it may not contain the elements inside.

Minimal complete definition

getRange, setRange

Methods

getRange :: a -> SrcSpan

setRange :: SrcSpan -> a -> a

annListElems :: RefMonads w r => Reference w r MU MU (AnnListG elem dom stage) (AnnListG elem dom stage) [Ann elem dom stage] [Ann elem dom stage]

annListAnnot :: RefMonads w r => Reference w r MU MU (AnnListG elem dom stage) (AnnListG elem dom stage) (NodeInfo (SemanticInfo dom (AnnListG elem)) (ListInfo stage)) (NodeInfo (SemanticInfo dom (AnnListG elem)) (ListInfo stage))

annList :: (RefMonads w r, MonadPlus r, Morph Maybe r, Morph [] r) => Reference w r MU MU (AnnListG e d s) (AnnListG e d s) (Ann e d s) (Ann e d s)

annJust :: (Functor w, Applicative w, Monad w, Functor r, Applicative r, MonadPlus r, Morph Maybe r) => Reference w r MU MU (AnnMaybeG e d s) (AnnMaybeG e d s) (Ann e d s) (Ann e d s)

annMaybe :: RefMonads w r => Reference w r MU MU (AnnMaybeG elem dom stage) (AnnMaybeG elem dom stage) (Maybe (Ann elem dom stage)) (Maybe (Ann elem dom stage))

isAnnNothing :: AnnMaybeG e d s -> Bool

type Domain d = (Typeable d, Data d, SemanticInfo' d SemaInfoDefaultCls ~ NoSemanticInfo, Data (SemanticInfo' d SemaInfoNameCls), Data (SemanticInfo' d SemaInfoLitCls), Data (SemanticInfo' d SemaInfoExprCls), Data (SemanticInfo' d SemaInfoImportCls), Data (SemanticInfo' d SemaInfoModuleCls), Data (SemanticInfo' d SemaInfoWildcardCls))

A semantic domain for the AST. The semantic domain maps semantic information for the different types of nodes in the AST. The kind of semantic domain for an AST depends on which stages of the compilation it passed. However after transforming the GHC representation to our AST, the domain stays the same. The domain is not applied to the AST elements that are generated while refactoring.

data Dom name

With this domain, semantic information can be parameterized. In practice it is only used if the compilation cannot proceed past the type checking phase.

data IdDom

shortShowSpan :: SrcSpan -> String

A short form of showing a range, without file name, for debugging purposes.

shortShowSpanWithFile :: SrcSpan -> String

data SrcTemplateStage

A stage where the annotation controls how the original source code can be retrieved from the AST. A source template is assigned to each node. It has holes where the content of an other node should be printed and ranges for the source code of the node.

class SourceInfoTraversal (a :: * -> * -> *) where

A class for traversing source information in an AST

Minimal complete definition

sourceInfoTraverseUp, sourceInfoTraverseDown, sourceInfoTraverse

Methods

sourceInfoTraverseUp :: Monad f => SourceInfoTrf f st1 st2 -> f () -> f () -> a dom st1 -> f (a dom st2)

sourceInfoTraverseDown :: Monad f => SourceInfoTrf f st1 st2 -> f () -> f () -> a dom st1 -> f (a dom st2)

sourceInfoTraverse :: Monad f => SourceInfoTrf f st1 st2 -> a dom st1 -> f (a dom st2)

sourceTemplateNodeRange :: Simple Lens (SpanInfo SrcTemplateStage) SrcSpan

sourceTemplateNodeElems :: Simple Lens (SpanInfo SrcTemplateStage) [SourceTemplateElem]

sourceTemplateListRange :: Simple Lens (ListInfo SrcTemplateStage) SrcSpan

srcTmpListBefore :: Simple Lens (ListInfo SrcTemplateStage) String

srcTmpListAfter :: Simple Lens (ListInfo SrcTemplateStage) String

srcTmpDefaultSeparator :: Simple Lens (ListInfo SrcTemplateStage) String

srcTmpIndented :: Simple Lens (ListInfo SrcTemplateStage) (Maybe [Bool])

srcTmpSeparators :: Simple Lens (ListInfo SrcTemplateStage) [([SourceTemplateTextElem], SrcSpan)]

sourceTemplateOptRange :: Simple Lens (OptionalInfo SrcTemplateStage) SrcSpan

srcTmpOptBefore :: Simple Lens (OptionalInfo SrcTemplateStage) String

srcTmpOptAfter :: Simple Lens (OptionalInfo SrcTemplateStage) String

data SourceTemplateTextElem

Constructors

NormalText

Fields

_sourceTemplateText :: String

StayingText

Fields

_sourceTemplateText :: String

_lineEndings :: String

sourceTemplateText :: Lens SourceTemplateTextElem SourceTemplateTextElem String String

data UnsupportedExtension

Constructors

UnsupportedExtension String

data SpliceInsertionProblem

Constructors

SpliceInsertionProblem SrcSpan String

data ConvertionProblem

Constructors

ConvertionProblem SrcSpan String

UnrootedConvertionProblem String

data TransformationProblem

Constructors

TransformationProblem String

data BreakUpProblem

Constructors

BreakUpProblem

Fields

bupOuter :: RealSrcSpan

bupInner :: SrcSpan

bupSiblings :: [SrcSpan]

data PrettyPrintProblem

Constructors

PrettyPrintProblem String