Language.Haskell.Tools.Refactor.Refactoring

Description

Defines a representation to represent refactorings that can be executed on the codebase. Refactorings are differentiated on their signatures (inputs needed to execute).

Documentation

data RefactoringChoice

The signature and behavior of one refactoring that can be executed.

Constructors

NamingRefactoring

Fields

refactoringName :: String

namingRefactoring :: RealSrcSpan -> String -> Refactoring

NamingRefactoringIndent

Fields

refactoringName :: String

namingRefactoringIndent :: RealSrcSpan -> String -> Maybe String -> Refactoring

SelectionRefactoring

Fields

refactoringName :: String

selectionRefactoring :: RealSrcSpan -> Refactoring

ModuleRefactoring

Fields

refactoringName :: String

moduleRefactoring :: Refactoring

ProjectRefactoring

Fields

refactoringName :: String

projectRefactoring :: ProjectRefactoring

performCommand

Arguments

:: [RefactoringChoice]
The set of available refactorings

-> [String]
The refactoring command

-> Either FilePath ModuleDom
The module in which the refactoring is performed

-> [ModuleDom]
Other modules

-> Ghc (Either String [RefactorChange])

Executes a given command (choosen from the set of available refactorings) on the selected module and given other modules.

refactorCommands :: [RefactoringChoice] -> [String]

Gets the name of possible refactorings.