Language.Haskell.Tools.Refactor.Utils.AST

Description

Operations for changing the AST

Documentation

removeChild :: SourceInfoTraversal e => e dom SrcTemplateStage -> LocalRefactor ()

Remove an element from the AST while keeping the textual parts of it that should not be removed (like preprocessor pragmas).

removeSeparator :: ([SourceTemplateTextElem], SrcSpan) -> LocalRefactor ()

Remove a separator from the AST while keeping the textual parts of it that should not be removed (like preprocessor pragmas).