Language.Haskell.Tools.Refactor.Utils.Helpers

Description

Helper functions for defining refactorings.

Documentation

replaceWithJust :: Ann e IdDom SrcTemplateStage -> AnnMaybe e -> AnnMaybe e

replaceWithNothing :: AnnMaybe e -> AnnMaybe e

removeEmptyBnds :: Simple Traversal Module ValueBind -> Simple Traversal Module Expr -> Module -> Module

Remove the container (where or let) when the last binding is removed.

normalizeElements :: [Ann e dom SrcTemplateStage] -> [Ann e dom SrcTemplateStage]

Puts the elements in the orginal order and remove duplicates (elements with the same source range)

groupElemsBy :: Ord k => (a -> k) -> [a] -> [[a]]

Groups elements together into equivalence groups.

reprElems :: [[a]] -> [(a, [a])]

Chooses a representative element for each equivalence group, and pairs them with their corresponding group.

equivalenceGroupsBy :: Ord k => (a -> k) -> [a] -> [(a, [a])]

Sorts the elements of a list into equivalence groups based on a function, then chooses a representative element for each group, and pairs them with their corresponding group.