Language.Haskell.Tools.Refactor.Utils.Maybe

Documentation

isJustT :: Monad m => MaybeT m a -> m Bool

isNothingT :: Monad m => MaybeT m a -> m Bool

liftMaybe :: Monad m => Maybe a -> MaybeT m a

fromMaybeT :: Monad m => a -> MaybeT m a -> m a

fromMaybeTM :: Monad m => m a -> MaybeT m a -> m a

maybeT :: Monad m => b -> (a -> b) -> MaybeT m a -> m b

maybeTM :: Monad m => m b -> (a -> m b) -> MaybeT m a -> m b

module Data.Maybe

newtype MaybeT (m :: Type -> Type) a

The parameterizable maybe monad, obtained by composing an arbitrary monad with the Maybe monad.

Computations are actions that may produce a value or exit.

The return function yields a computation that produces that value, while >>= sequences two subcomputations, exiting if either computation does.

Constructors

MaybeT

Fields

runMaybeT :: m (Maybe a)