Language.Haskell.Tools.Refactor.Utils.NameLookup

Documentation

opSemName :: Operator -> Maybe Name

declHeadQName :: DeclHead -> QualifiedName

declHeadSemName :: DeclHead -> Maybe Name

instHeadSemName :: InstanceHead -> Maybe Name

assertionQNames :: Assertion -> [QualifiedName]

Collects the qualified names of the class heads in an assertion.

assertionSemNames :: Assertion -> [Name]

Collects the semantic names of the class heads in an assertion.

nameFromType :: Type -> Maybe Name

Extracts the name of a type. In case of a type application, it finds the type being applied. It works only for unambiguous types, so it won't work for tuples.

Orphan instances

HasNameInfo' Name

Instance details

Methods

semanticsName :: Name -> Maybe Name

HasNameInfo' DeclHead

Instance details

Methods

semanticsName :: DeclHead -> Maybe Name

HasNameInfo' InstanceHead

Instance details

Methods

semanticsName :: InstanceHead -> Maybe Name

HasNameInfo' Operator

Instance details

Methods

semanticsName :: Operator -> Maybe Name