Language.Haskell.Tools.Refactor

Description

Defines the API for refactorings

Documentation

module Language.Haskell.Tools.AST.SemaInfoClasses

module Language.Haskell.Tools.Rewrite

module Language.Haskell.Tools.AST.References

module Language.Haskell.Tools.AST.Helpers

module Language.Haskell.Tools.Refactor.Utils.Monadic

module Language.Haskell.Tools.Refactor.Utils.Helpers

module Language.Haskell.Tools.Rewrite.ElementTypes

module Language.Haskell.Tools.Refactor.Prepare

module Language.Haskell.Tools.Refactor.Utils.Lists

module Language.Haskell.Tools.Refactor.Utils.BindingElem

module Language.Haskell.Tools.Refactor.Utils.Indentation

module Language.Haskell.Tools.Refactor.Refactoring

module Language.Haskell.Tools.Refactor.Utils.Name

module Language.Haskell.Tools.Refactor.Representation

module Language.Haskell.Tools.Refactor.Monad

data Ann (elem :: * -> * -> *) dom stage :: (* -> * -> *) -> * -> * -> *

An element of the AST keeping extra information.

class HasSourceInfo e where

Minimal complete definition

srcInfo

Associated Types

type SourceInfoType e :: *

Methods

srcInfo :: Simple Lens e (SourceInfoType e)

class HasRange a where

Extracts or modifies the concrete range corresponding to a given source info. In case of lists and optional elements, it may not contain the elements inside.

Minimal complete definition

getRange, setRange

Methods

getRange :: a -> SrcSpan

setRange :: SrcSpan -> a -> a

annListElems :: RefMonads w r => Reference w r MU MU (AnnListG elem dom stage) (AnnListG elem dom stage) [Ann elem dom stage] [Ann elem dom stage]

annListAnnot :: RefMonads w r => Reference w r MU MU (AnnListG elem dom stage) (AnnListG elem dom stage) (NodeInfo (SemanticInfo dom (AnnListG elem)) (ListInfo stage)) (NodeInfo (SemanticInfo dom (AnnListG elem)) (ListInfo stage))

annList :: (RefMonads w r, MonadPlus r, Morph Maybe r, Morph [] r) => Reference w r MU MU (AnnListG e d s) (AnnListG e d s) (Ann e d s) (Ann e d s)

annJust :: (Functor w, Applicative w, Monad w, Functor r, Applicative r, MonadPlus r, Morph Maybe r) => Reference w r MU MU (AnnMaybeG e d s) (AnnMaybeG e d s) (Ann e d s) (Ann e d s)

annMaybe :: RefMonads w r => Reference w r MU MU (AnnMaybeG elem dom stage) (AnnMaybeG elem dom stage) (Maybe (Ann elem dom stage)) (Maybe (Ann elem dom stage))

isAnnNothing :: AnnMaybeG e d s -> Bool

type Domain d = (Typeable * d, Data d, (~) * (SemanticInfo' d SameInfoDefaultCls) NoSemanticInfo, Data (SemanticInfo' d SameInfoNameCls), Data (SemanticInfo' d SameInfoExprCls), Data (SemanticInfo' d SameInfoImportCls), Data (SemanticInfo' d SameInfoModuleCls), Data (SemanticInfo' d SameInfoWildcardCls), Show (SemanticInfo' d SameInfoNameCls), Show (SemanticInfo' d SameInfoExprCls), Show (SemanticInfo' d SameInfoImportCls), Show (SemanticInfo' d SameInfoModuleCls), Show (SemanticInfo' d SameInfoWildcardCls))

A semantic domain for the AST. The semantic domain maps semantic information for the different types of nodes in the AST. The kind of semantic domain for an AST depends on which stages of the compilation did it pass. However after transforming the GHC representation to our AST, the domain keeps the same. The domain is not applied to the AST elements that are generated while refactoring.

data Dom name :: * -> *

With this domain, semantic information can be parameterized. In practice it is only used if the compilation cannot proceed past the type checking phase.

data IdDom :: *

shortShowSpan :: SrcSpan -> String

A short form of showing a range, without file name, for debugging purposes.

shortShowSpanWithFile :: SrcSpan -> String

data SrcTemplateStage :: *

A stage where the annotation controls how the original source code can be retrieved from the AST. A source template is assigned to each node. It has holes where the content of an other node should be printed and ranges for the source code of the node.

class SourceInfoTraversal (a :: * -> * -> *) where

A class for traversing source information in an AST

Minimal complete definition

sourceInfoTraverseUp, sourceInfoTraverseDown, sourceInfoTraverse

Methods

sourceInfoTraverseUp :: Monad f => SourceInfoTrf f st1 st2 -> f () -> f () -> a dom st1 -> f (a dom st2)

sourceInfoTraverseDown :: Monad f => SourceInfoTrf f st1 st2 -> f () -> f () -> a dom st1 -> f (a dom st2)

sourceInfoTraverse :: Monad f => SourceInfoTrf f st1 st2 -> a dom st1 -> f (a dom st2)

sourceTemplateNodeRange :: Simple Lens (SpanInfo SrcTemplateStage) SrcSpan

sourceTemplateNodeElems :: Simple Lens (SpanInfo SrcTemplateStage) [SourceTemplateElem]

sourceTemplateListRange :: Simple Lens (ListInfo SrcTemplateStage) SrcSpan

srcTmpListBefore :: Simple Lens (ListInfo SrcTemplateStage) String

srcTmpListAfter :: Simple Lens (ListInfo SrcTemplateStage) String

srcTmpDefaultSeparator :: Simple Lens (ListInfo SrcTemplateStage) String

srcTmpIndented :: Simple Lens (ListInfo SrcTemplateStage) (Maybe [Bool])

srcTmpSeparators :: Simple Lens (ListInfo SrcTemplateStage) [([SourceTemplateTextElem], SrcSpan)]

sourceTemplateOptRange :: Simple Lens (OptionalInfo SrcTemplateStage) SrcSpan

srcTmpOptBefore :: Simple Lens (OptionalInfo SrcTemplateStage) String

srcTmpOptAfter :: Simple Lens (OptionalInfo SrcTemplateStage) String

data SourceTemplateTextElem :: *

Constructors

NormalText

Fields

_sourceTemplateText :: String

StayingText

Fields

_sourceTemplateText :: String

_lineEndings :: String

sourceTemplateText :: Lens SourceTemplateTextElem SourceTemplateTextElem String String

data UnsupportedExtension

Constructors

UnsupportedExtension String

data SpliceInsertionProblem :: *

Constructors

SpliceInsertionProblem SrcSpan String

data ConvertionProblem :: *

Constructors

ConvertionProblem SrcSpan String

UnrootedConvertionProblem String

data TransformationProblem :: *

Constructors

TransformationProblem String

data BreakUpProblem :: *

Constructors

BreakUpProblem

Fields

bupOuter :: RealSrcSpan

bupInner :: SrcSpan

bupSiblings :: [SrcSpan]

data PrettyPrintProblem :: *

Constructors

PrettyPrintProblem String