Language.Haskell.Tools.Refactor.Monad

Description

Types and instances for monadic refactorings. The refactoring monad provides automatic importing, keeping important source fragments (such as preprocessor pragmas), and providing contextual information for refactorings.

Documentation

class Monad m => RefactorMonad m where

A monad that can be used to refactor

Minimal complete definition

refactError, liftGhc

Methods

refactError :: String -> m a

liftGhc :: Ghc a -> m a

type LocalRefactoring = UnnamedModule -> LocalRefactor UnnamedModule

A refactoring that only affects one module

type Refactoring = ModuleDom -> [ModuleDom] -> Refactor [RefactorChange]

The type of a refactoring

type ProjectRefactoring = [ModuleDom] -> Refactor [RefactorChange]

The type of a refactoring that affects the whole project.

type LocalRefactor = LocalRefactorT Refactor

The refactoring monad for a given module

type Refactor = ExceptT String Ghc

The refactoring monad for the whole project

newtype LocalRefactorT m a

Input and output information for the refactoring TODO: use multiple states instead of Either

Constructors

LocalRefactorT

Fields

fromRefactorT :: WriterT [Either Name (SrcSpan, String, String)] (ReaderT RefactorCtx m) a

data RefactorCtx

The information a refactoring can use

Constructors

RefactorCtx

Fields

refModuleName :: Module

The name of the module being refactored. Used for accessing implicit imports.

refCtxRoot :: Ann UModule IdDom SrcTemplateStage

refCtxImports :: [Ann UImportDecl IdDom SrcTemplateStage]

Some instances missing from GHC

Orphan instances

ExceptionMonad m => ExceptionMonad (MaybeT m)

Methods

gcatch :: Exception e => MaybeT m a -> (e -> MaybeT m a) -> MaybeT m a

gmask :: ((MaybeT m a -> MaybeT m a) -> MaybeT m b) -> MaybeT m b

gbracket :: MaybeT m a -> (a -> MaybeT m b) -> (a -> MaybeT m c) -> MaybeT m c

gfinally :: MaybeT m a -> MaybeT m b -> MaybeT m a

(GhcMonad m, Monoid s) => GhcMonad (WriterT s m)

Methods

getSession :: WriterT s m HscEnv

setSession :: HscEnv -> WriterT s m ()

GhcMonad m => GhcMonad (StateT s m)

Methods

getSession :: StateT s m HscEnv

setSession :: HscEnv -> StateT s m ()

GhcMonad m => GhcMonad (StateT s m)

Methods

getSession :: StateT s m HscEnv

setSession :: HscEnv -> StateT s m ()

GhcMonad m => GhcMonad (ExceptT s m)

Methods

getSession :: ExceptT s m HscEnv

setSession :: HscEnv -> ExceptT s m ()

(Monad m, HasDynFlags m) => HasDynFlags (StateT s m)

Methods

getDynFlags :: StateT s m DynFlags

(Monad m, HasDynFlags m) => HasDynFlags (StateT s m)

Methods

getDynFlags :: StateT s m DynFlags

(ExceptionMonad m, Monoid s) => ExceptionMonad (WriterT s m)

Methods

gcatch :: Exception e => WriterT s m a -> (e -> WriterT s m a) -> WriterT s m a

gmask :: ((WriterT s m a -> WriterT s m a) -> WriterT s m b) -> WriterT s m b

gbracket :: WriterT s m a -> (a -> WriterT s m b) -> (a -> WriterT s m c) -> WriterT s m c

gfinally :: WriterT s m a -> WriterT s m b -> WriterT s m a

ExceptionMonad m => ExceptionMonad (StateT s m)

Methods

gcatch :: Exception e => StateT s m a -> (e -> StateT s m a) -> StateT s m a

gmask :: ((StateT s m a -> StateT s m a) -> StateT s m b) -> StateT s m b

gbracket :: StateT s m a -> (a -> StateT s m b) -> (a -> StateT s m c) -> StateT s m c

gfinally :: StateT s m a -> StateT s m b -> StateT s m a

ExceptionMonad m => ExceptionMonad (StateT s m)

Methods

gcatch :: Exception e => StateT s m a -> (e -> StateT s m a) -> StateT s m a

gmask :: ((StateT s m a -> StateT s m a) -> StateT s m b) -> StateT s m b

gbracket :: StateT s m a -> (a -> StateT s m b) -> (a -> StateT s m c) -> StateT s m c

gfinally :: StateT s m a -> StateT s m b -> StateT s m a

ExceptionMonad m => ExceptionMonad (ExceptT s m)

Methods

gcatch :: Exception e => ExceptT s m a -> (e -> ExceptT s m a) -> ExceptT s m a

gmask :: ((ExceptT s m a -> ExceptT s m a) -> ExceptT s m b) -> ExceptT s m b

gbracket :: ExceptT s m a -> (a -> ExceptT s m b) -> (a -> ExceptT s m c) -> ExceptT s m c

gfinally :: ExceptT s m a -> ExceptT s m b -> ExceptT s m a

GhcMonad m => GhcMonad (ReaderT * s m)

Methods

getSession :: ReaderT * s m HscEnv

setSession :: HscEnv -> ReaderT * s m ()

ExceptionMonad m => ExceptionMonad (ReaderT * s m)

Methods

gcatch :: Exception e => ReaderT * s m a -> (e -> ReaderT * s m a) -> ReaderT * s m a

gmask :: ((ReaderT * s m a -> ReaderT * s m a) -> ReaderT * s m b) -> ReaderT * s m b

gbracket :: ReaderT * s m a -> (a -> ReaderT * s m b) -> (a -> ReaderT * s m c) -> ReaderT * s m c

gfinally :: ReaderT * s m a -> ReaderT * s m b -> ReaderT * s m a