Language.Haskell.Tools.Refactor.Prepare

Description

Defines utility methods that prepare Haskell modules for refactoring

Documentation

tryRefactor :: (RealSrcSpan -> Refactoring) -> String -> String -> IO ()

A quick function to try the refactorings

correctRefactorSpan :: UnnamedModule -> RealSrcSpan -> RealSrcSpan

Adjust the source range to be applied to the refactored module

useFlags :: [String] -> Ghc ([String], DynFlags -> DynFlags)

Set the given flags for the GHC session. Also gives back a change function that you can use to apply the settings to any flags. Prints out errors and warnings

reloadPkgDb :: Ghc ()

Reloads the package database based on the session flags

initGhcFlags :: Ghc ()

Initialize GHC flags to default values that support refactoring

initGhcFlagsForTest :: Ghc ()

initGhcFlags' :: Bool -> Ghc ()

Sets up basic flags and settings for GHC

useDirs :: [FilePath] -> Ghc ()

Use the given source directories when searching for imported modules

deregisterDirs :: [FilePath] -> Ghc ()

Don't use the given source directories when searching for imported modules

toFileName :: FilePath -> String -> FilePath

Translates module name and working directory into the name of the file where the given module should be defined

toBootFileName :: FilePath -> String -> FilePath

Translates module name and working directory into the name of the file where the boot module should be defined

getSourceDir :: ModSummary -> IO FilePath

Get the source directory where the module is located.

getModSumOrig :: ModSummary -> FilePath

Gets the path to the source file of the module.

keyFromMS :: ModSummary -> SourceFileKey

getModSumName :: ModSummary -> String

Gets the module name

loadModule :: String -> String -> Ghc ModSummary

Load the summary of a module given by the working directory and module name.

type TypedModule = Ann UModule IdDom SrcTemplateStage

The final version of our AST, with type infromation added

parseTyped :: ModSummary -> Ghc TypedModule

Get the typed representation of a Haskell module.

data UnsupportedExtension

Constructors

UnsupportedExtension String

trfProblem :: String -> a

withAlteredDynFlags :: GhcMonad m => (DynFlags -> m DynFlags) -> m a -> m a

Modifies the dynamic flags for performing a ghc task

forceCodeGen :: ModSummary -> ModSummary

Forces the code generation for a given module

codeGenDfs :: DynFlags -> DynFlags

forceAsmGen :: ModSummary -> ModSummary

Forces ASM code generation for a given module

modSumNormalizeFlags :: ModSummary -> ModSummary

Normalizes the flags for a module summary

normalizeFlags :: DynFlags -> DynFlags

Removes all flags that are unintelligable for refactoring

readSrcSpan :: String -> RealSrcSpan

Read a source range from our textual format: line:col-line:col or line:col

readSrcLoc :: String -> RealSrcLoc

Read a source location from our format: line:col