Language.Haskell.Tools.Refactor.Representation

Description

Representation of modules, their collections, refactoring changes and exceptions.

Documentation

type UnnamedModule dom = Ann UModule dom SrcTemplateStage

A type for the input and result of refactoring a module

type ModuleDom dom = (SourceFileKey, UnnamedModule dom)

The name of the module and the AST

data SourceFileKey

Module name and marker to separate .hs-boot module definitions. Specifies a source file in a working directory.

Constructors

SourceFileKey

Fields

_sfkFileName :: FilePath

_sfkModuleName :: String

data RefactorChange dom

Change in the project, modification or removal of a module.

Constructors

ContentChanged

Fields

fromContentChanged :: ModuleDom dom

ModuleRemoved

Fields

removedModuleName :: String

ModuleCreated

Fields

createdModuleName :: String

createdModuleContent :: UnnamedModule dom

sameLocation :: SourceFileKey

data RefactorException

Exceptions that can occur while loading modules or during internal operations (not during performing the refactor).

Constructors

IllegalExtensions [String]

SourceCodeProblem ErrorMessages

UnknownException String

moduleSourceFile :: String -> FilePath

Transforms module name to a .hs file name relative to the source root directory.

sourceFileModule :: FilePath -> String

Transforms a source root relative file name into module name.

sfkModuleName :: Lens SourceFileKey SourceFileKey String String

sfkFileName :: Lens SourceFileKey SourceFileKey FilePath FilePath

Orphan instances

Show ErrorMessages

Methods

showsPrec :: Int -> ErrorMessages -> ShowS

show :: ErrorMessages -> String

showList :: [ErrorMessages] -> ShowS