Language.Haskell.Tools.Refactor.Utils.BindingElem

Description

Utilities for transformations that work on both top-level and local definitions

Documentation

class NamedElement d => BindingElem d where

A type class for handling definitions that can appear as both top-level and local definitions

Minimal complete definition

sigBind, valBind, fixitySig, createTypeSig, createBinding, createFixitySig, isTypeSig, isBinding, isFixitySig

Methods

sigBind :: Simple Partial (Ann d IdDom SrcTemplateStage) TypeSignature

Accesses a type signature definition in a local or top-level definition

valBind :: Simple Partial (Ann d IdDom SrcTemplateStage) ValueBind

Accesses a value or function definition in a local or top-level definition

fixitySig :: Simple Partial (Ann d IdDom SrcTemplateStage) FixitySignature

Accesses a type signature definition in a local or top-level definition

createTypeSig :: TypeSignature -> Ann d IdDom SrcTemplateStage

Creates a new definition from a type signature

createBinding :: ValueBind -> Ann d IdDom SrcTemplateStage

Creates a new definition from a value or function definition

createFixitySig :: FixitySignature -> Ann d IdDom SrcTemplateStage

Creates a new fixity signature

isTypeSig :: Ann d IdDom SrcTemplateStage -> Bool

Checks if a given definition is a type signature

isBinding :: Ann d IdDom SrcTemplateStage -> Bool

Checks if a given definition is a function or value binding

isFixitySig :: Ann d IdDom SrcTemplateStage -> Bool

Checks if a given definition is a fixity signature

getValBindInList :: BindingElem d => RealSrcSpan -> AnnList d -> Maybe ValueBind

valBindsInList :: BindingElem d => Simple Traversal (AnnList d) ValueBind