Language.Haskell.Tools.Refactor.Utils.Lists

Description

Defines operation on AST lists. AST lists carry source information so simple list modification is not enough.

Documentation

filterList :: SourceInfoTraversal e => (Ann e dom SrcTemplateStage -> Bool) -> AnnList e dom -> AnnList e dom

Filters the elements of the list. By default it removes the separator before the element. Of course, if the first element is removed, the following separator is removed as well.

filterListIndexed :: SourceInfoTraversal e => (Int -> Ann e dom SrcTemplateStage -> Bool) -> AnnList e dom -> AnnList e dom

filterListSt :: SourceInfoTraversal e => (Ann e dom SrcTemplateStage -> Bool) -> AnnList e dom -> LocalRefactor dom (AnnList e dom)

A version of filterList that cares about keeping non-removable code elements (like preprocessor pragmas)

filterListIndexedSt :: SourceInfoTraversal e => (Int -> Ann e dom SrcTemplateStage -> Bool) -> AnnList e dom -> LocalRefactor dom (AnnList e dom)

A version of filterListIndexed that cares about keeping non-removable code elements (like preprocessor pragmas)

sublist :: [Int] -> [a] -> [a]

Selects the given indices from a list

notSublist :: [Int] -> [a] -> [a]

Selects all but the given indices from a list

insertWhere :: Bool -> Ann e dom SrcTemplateStage -> (Maybe (Ann e dom SrcTemplateStage) -> Bool) -> (Maybe (Ann e dom SrcTemplateStage) -> Bool) -> AnnList e dom -> AnnList e dom

Inserts the element in the places where the two positioning functions (one checks the element before, one the element after) allows the placement.

insertIndex :: (Maybe (Ann e dom SrcTemplateStage) -> Bool) -> (Maybe (Ann e dom SrcTemplateStage) -> Bool) -> [Ann e dom SrcTemplateStage] -> Maybe Int

Checks where the element will be inserted given the two positioning functions.

zipWithSeparators :: AnnList e dom -> [(([SourceTemplateTextElem], SrcSpan), Ann e dom SrcTemplateStage)]

Gets the elements and separators from a list. The first separator is zipped to the second element. To the first element, the "" string is zipped.