Language.Haskell.Tools.Refactor.Utils.Maybe

Documentation

liftMaybe :: Monad m => Maybe a -> MaybeT m a

maybeT :: Monad m => b -> (a -> b) -> MaybeT m a -> m b

maybeTM :: Monad m => m b -> (a -> m b) -> MaybeT m a -> m b

module Data.Maybe

newtype MaybeT (m :: * -> *) a :: (* -> *) -> * -> *

The parameterizable maybe monad, obtained by composing an arbitrary monad with the Maybe monad.

Computations are actions that may produce a value or exit.

The return function yields a computation that produces that value, while >>= sequences two subcomputations, exiting if either computation does.

Constructors

MaybeT

Fields

runMaybeT :: m (Maybe a)